requirements.txt
psycopg2-binary
chromadb
python-dotenv
requests
pyalex
sqlalchemy